feat(app-platform): App Auth[entication|orization]
Some requests made by a SentryApp will use the Client ID/Secret as
authentication. For example, the request to exchange a Grant Code for an
Access Token. This adds logic to the Authentication middleware to find
the User based on that data.

Also modifies the Access granting logic to be SentryApp-aware. When a
request is made as a SentryApp (using an Access Token they exchanged
for), we grant access to Organizations, and Teams, it's installed on.

Also adds a few Endpoint/Permission base classes for future Endpoint
work.