feat(gitlab) Implement push & merge_request hooks for gitlab
Tweak how we handle gitlab external_id and webhook secrets. GitLab push
events do not provide much context for us to locate the sentry
organization and repository. We need to leverage the only field we have
(the webhook secret) to carry an identifier that lets us locate the
integration installation. With this and the rest of the push payload we
can locate the repository to sync commits in.

This does not handle pushes that contain more than 20 commits. Like
GitHub we only sync the first 20 commits of any push. The remaining diff
gets collected when a Release is created.

Refs APP-597